fix: 排课页 null start_time 崩溃 + duration 字段支持
- Schedule.jsx: getAvailableTeachers/getSchedulesForSlot 加 null 防御
- Schedule.jsx: handleEditSchedule 优先用 schedule.duration 回显
- Schedule.jsx: 修复 end_time 计算漏掉分钟进位
- classes.js: INSERT 加 duration 列 + 修复 cls 未定义 bug
- validation.js: classSchema 新增 duration 字段
- org-portal: 同步修复 50min 课计算逻辑
- wrangler.toml: DEBUG=true (调试期间开启)

+++ api/src/routes/students.js
@@ -353,7 +353,7 @@ students.patch('/:id/add-hours', validateParams(idParamSchema), async (c) => {
   const DB = c.env.DB;
   const { id } = c.req.validatedParams;
   const body = await c.req.json();
-  const hours = parseInt(body.hours) || 0;
+  const hours = parseFloat(body.hours) || 0;
 
   if (hours <= 0) {
     return c.json(error('BAD_REQUEST', '课时数必须大于0'), 400);
